requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==1.0.1
